feat:added h2-console bean in beans.xml
refactor:changed cgpa field from double to BigDecimal

feat:created schema.sql file for db initialization

+++ src/main/java/com/studenthub/App.java
@@ -1,5 +1,7 @@
 package com.studenthub;
 
+import java.math.BigDecimal;
+
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
@@ -11,7 +13,7 @@ public class App
     {
         ApplicationContext context = new ClassPathXmlApplicationContext("beans.xml");
         System.out.println("application context initialized");
-        StudentDto dto = new StudentDto(1,"hari","chn",7.89,001,"CSE");
+        StudentDto dto = new StudentDto(1,"hari","chn",BigDecimal.valueOf(7.89),001,"CSE");
         System.out.println(dto);
         ((ClassPathXmlApplicationContext)context).close();
         System.out.println("application context closed");
